Function makair_telemetry::gather_telemetry

pub fn gather_telemetry(
    port_id: &str, 
    tx: Sender<TelemetryChannelType>, 
    file_buf: Option<BufWriter<File>>, 
    control_rx: Option<Receiver<ControlMessage>>
)

This is supported on crate feature serial only.

Open a serial port, consume it endlessly and send parsed telemetry messages through a channel

port_id - Name or path to the serial port.

tx - Sender of a channel.

file_buf - Optional file buffer; if specified, messages will also be serialized and written in this file.

control_rx - Optional receiver of a channel used to send control messages through the serial port.

This is meant to be run in a dedicated thread.