Struct makair_telemetry::structures::MachineStateSnapshot

pub struct MachineStateSnapshot {

    pub telemetry_version: u8,
    pub version: String,
    pub device_id: String,
    pub systick: u64,
    pub cycle: u32,
    pub peak_command: u8,
    pub plateau_command: u8,
    pub peep_command: u8,
    pub cpm_command: u8,
    pub previous_peak_pressure: u16,
    pub previous_plateau_pressure: u16,
    pub previous_peep_pressure: u16,
    pub current_alarm_codes: Vec<u8>,
    pub previous_volume: Option<u16>,
    pub expiratory_term: u8,
    pub trigger_enabled: bool,
    pub trigger_offset: u8,
    pub previous_cpm: Option<u8>,
    pub alarm_snoozed: Option<bool>,
    pub cpu_load: Option<u8>,
    pub ventilation_mode: VentilationMode,
    pub inspiratory_trigger_flow: Option<u8>,
    pub expiratory_trigger_flow: Option<u8>,
    pub ti_min: Option<u16>,
    pub ti_max: Option<u16>,
    pub low_inspiratory_minute_volume_alarm_threshold: Option<u8>,
    pub high_inspiratory_minute_volume_alarm_threshold: Option<u8>,
    pub low_expiratory_minute_volume_alarm_threshold: Option<u8>,
    pub high_expiratory_minute_volume_alarm_threshold: Option<u8>,
    pub low_respiratory_rate_alarm_threshold: Option<u8>,
    pub high_respiratory_rate_alarm_threshold: Option<u8>,
    pub target_tidal_volume: Option<u16>,
    pub low_tidal_volume_alarm_threshold: Option<u16>,
    pub high_tidal_volume_alarm_threshold: Option<u16>,
    pub plateau_duration: Option<u16>,
    pub leak_alarm_threshold: Option<u16>,
    pub target_inspiratory_flow: Option<u8>,
    pub inspiratory_duration_command: Option<u16>,
    pub previous_inspiratory_duration: Option<u16>,
    pub battery_level: Option<u16>,
    pub locale: Option<Locale>,
    pub patient_height: Option<u8>,
    pub patient_gender: Option<PatientGender>,
    pub peak_pressure_alarm_threshold: Option<u16>,

}

A telemetry message that is sent at the end of every respiratory cycle

Fields
telemetry_version: u8

Version of the telemetry protocol

version: String

Version of the MCU firmware

device_id: String

Internal ID of the MCU

systick: u64

Number of microseconds since the MCU booted

cycle: u32

Number of the current breathing cycle since MCU booted

peak_command: u8

Requested peak command in cmH2O

plateau_command: u8

Requested plateau command in cmH2O

peep_command: u8

Requested PEEP command in cmH2O

cpm_command: u8

Requested number of cycles per minute

previous_peak_pressure: u16

Measured peak pressure in mmH2O

previous_plateau_pressure: u16

Measured pleateau pressure in mmH2O

previous_peep_pressure: u16

Measured PEEP in mmH2O

current_alarm_codes: Vec<u8>

Codes of the alarms that are currently triggered

previous_volume: Option<u16>

Measured previous_volume in mL (sensor might not be enabled)

expiratory_term: u8

Expiration term in the “Inspiration/Expiration” ratio given that Inspiration = 10

trigger_enabled: bool

State of the trigger

trigger_offset: u8

Trigger offset in mmH2O

previous_cpm: Option<u8>

[protocol v2] Measured number of cycles per minute

alarm_snoozed: Option<bool>

[protocol v2] State of the alarm snooze

cpu_load: Option<u8>

[protocol v2] CPU load in percent

ventilation_mode: VentilationMode

Ventilation mode

inspiratory_trigger_flow: Option<u8>

[protocol v2] Inspiratory trigger flow in percent

expiratory_trigger_flow: Option<u8>

[protocol v2] Expiratory trigger flow in percent

ti_min: Option<u16>

[protocol v2] Minimum duration of inhalation in ms

ti_max: Option<u16>

[protocol v2] Maximum duration of inhalation in ms

low_inspiratory_minute_volume_alarm_threshold: Option<u8>

[protocol v2] Threshold for low inspiratory minute volume alarm in L/min

high_inspiratory_minute_volume_alarm_threshold: Option<u8>

[protocol v2] Threshold for high inspiratory minute volume alarm in L/min

low_expiratory_minute_volume_alarm_threshold: Option<u8>

[protocol v2] Threshold for low expiratory minute volume alarm in L/min

high_expiratory_minute_volume_alarm_threshold: Option<u8>

[protocol v2] Threshold for high expiratory minute volume alarm in L/min

low_respiratory_rate_alarm_threshold: Option<u8>

[protocol v2] Threshold for low respiratory rate alarm in cycle per minute

high_respiratory_rate_alarm_threshold: Option<u8>

[protocol v2] Threshold for high respiratory rate alarm in cycle per minute

target_tidal_volume: Option<u16>

[protocol v2] Target tidal volume in mL

low_tidal_volume_alarm_threshold: Option<u16>

[protocol v2] Threshold for low tidal volume in mL

high_tidal_volume_alarm_threshold: Option<u16>

[protocol v2] Threshold for high tidal volume in mL

plateau_duration: Option<u16>

[protocol v2] Duration in ms of closing both valves to effectively measure plateau pressure in volume control modes

leak_alarm_threshold: Option<u16>

[protocol v2] Threshold for leak alarm in cL/min

target_inspiratory_flow: Option<u8>

[protocol v2] Target flow during inspiration in L/min

inspiratory_duration_command: Option<u16>

[protocol v2] Requested duration of inspiration in ms

previous_inspiratory_duration: Option<u16>

[protocol v2] Measured duration of inspiration in ms

battery_level: Option<u16>

[protocol v2] Measured battery level value in centivolts (precise value)

locale: Option<Locale>

[protocol v2] Language of the system

patient_height: Option<u8>

[protocol v2] Patient’s height in centimeters

patient_gender: Option<PatientGender>

[protocol v2] Patient’s gender

peak_pressure_alarm_threshold: Option<u16>

[protocol v2] Threshold for peak pressure alarm in mmH2O