Crate makair_telemetry

This crate is a library that handles reading and parsing the MakAir’s telemetry binary protocol, and optionally sending new settings values using the Makair’s control binary protocol.

Re-exports

pub use serial;

pub use url;

Modules

alarm

Utilities related to alarms

control

Structures to represent control messages

error

Error-related entities

locale

Tools to manipulate ISO 639-1 language codes to be used in the control protocol

parsers

Underlying parsers for telemetry messages

serializers

Binary representation of telemtry messages

structures

Structures to represent telemetry messages

Functions

display_message

Helper to display telemetry messages

gather_telemetryserial

Open a serial port, consume it endlessly and send parsed telemetry messages through a channel

gather_telemetry_from_bytes

Open a byte channel, consume it endlessly and send parsed telemetry messages through another channel

gather_telemetry_from_file

Open a file containing serialized telemetry data, read it and send back parsed telemetry messages through a channel

gather_telemetry_from_wswebsocket

Connect to a WebSocket server, get binary messages endlessly and send parsed telemetry messages through a channel

Type Definitions

TelemetryChannelType

A decoded telemetry message