Enum makair_telemetry::control::ControlSetting

pub enum ControlSetting {

    Heartbeat,
    VentilationMode,
    PlateauPressure,
    PEEP,
    CyclesPerMinute,
    ExpiratoryTerm,
    TriggerEnabled,
    TriggerOffset,
    RespirationEnabled,
    AlarmSnooze,
    InspiratoryTriggerFlow,
    ExpiratoryTriggerFlow,
    TiMin,
    TiMax,
    LowInspiratoryMinuteVolumeAlarmThreshold,
    HighInspiratoryMinuteVolumeAlarmThreshold,
    LowExpiratoryMinuteVolumeAlarmThreshold,
    HighExpiratoryMinuteVolumeAlarmThreshold,
    LowRespiratoryRateAlarmThreshold,
    HighRespiratoryRateAlarmThreshold,
    TargetTidalVolume,
    LowTidalVolumeAlarmThreshold,
    HighTidalVolumeAlarmThreshold,
    PlateauDuration,
    LeakAlarmThreshold,
    TargetInspiratoryFlow,
    InspiratoryDuration,
    Locale,
    PatientHeight,
    PatientGender,
    PeakPressureAlarmThreshold,
    EolConfirm,

}

Available settings in the control protocol

Variants

Heartbeat

Heartbeat used for the RPi watchdog feature (value is ignored except for the special value DISABLE_RPI_WATCHDOG which disables watchdog)

VentilationMode

Ventilation mode, must be one of the following:

1 → PC-CMV

2 → PC-AC (default)

3 → VC-CMV

4 → PC-VSAI

5 → VC-AC

PlateauPressure

Plateau pressure in mmH2O (value bounds must be between 100 and 400)

PEEP

PEEP in mmH2O (value bounds must be between 0 and 300)

CyclesPerMinute

Number of cycles per minute (value bounds must be between 5 and 35)

ExpiratoryTerm

Expiration term in the “Inspiration/Expiration” ratio given that Inspiration = 10 (value bounds must be between 10 and 60)

TriggerEnabled

State of the trigger (value must be 1 if enabled and 0 if disabled)

TriggerOffset

Trigger offset in mmH2O (value bounds must be between 0 and 100)

RespirationEnabled

State of the respiration (value must be 1 if enabled and 0 if disabled)

AlarmSnooze

Alarm snooze (value must be 1 to snooze and 0 to unsnooze)

InspiratoryTriggerFlow

Inspiratory trigger flow in percent

ExpiratoryTriggerFlow

Expiratory trigger flow in percent

TiMin

Minimum duration of inhalation in ms (value bounds must be between 100 and 3000)

TiMax

Maximum duration of inhalation in ms (value bounds must be between 200 and 5000)

LowInspiratoryMinuteVolumeAlarmThreshold

Threshold for low inspiratory minute volume alarm in L/min (value bounds must be between 0 and 20)

HighInspiratoryMinuteVolumeAlarmThreshold

Threshold for high inspiratory minute volume alarm in L/min (value bounds must be between 10 and 40)

LowExpiratoryMinuteVolumeAlarmThreshold

Threshold for low expiratory minute volume alarm in L/min (value bounds must be between 0 and 20)

HighExpiratoryMinuteVolumeAlarmThreshold

Threshold for high expiratory minute volume alarm in L/min (value bounds must be between 10 and 40)

LowRespiratoryRateAlarmThreshold

Threshold for low respiratory rate alarm in cycle per minute (value bounds must be between 5 and 25)

HighRespiratoryRateAlarmThreshold

Threshold for high respiratory rate alarm in cycle per minute (value bounds must be between 15 and 35)

TargetTidalVolume

Target tidal volume in mL (value bounds must be between 50 and 2000)

LowTidalVolumeAlarmThreshold

Threshold for low tidal volume in mL (value bounds must be between 0 and 1000)

HighTidalVolumeAlarmThreshold

Threshold for high tidal volume in mL (value bounds must be between 50 and 2000)

PlateauDuration

Duration in ms of closing both valves to effectively measure plateau pressure in volume control modes (value bounds must be between 100 and 2000)

LeakAlarmThreshold

Threshold for leak alarm in cL/min (value bounds must be between 0 and 10000)

TargetInspiratoryFlow

Target flow during inspiration in L/min (value bounds must be between 5 and 80)

InspiratoryDuration

Duration of inspiration in ms (value bounds must be between 200 and 3000)

Locale

Language of the system; this should be two letters (see ISO 639-1) in ASCII representation as two u8

PatientHeight

Patient’s height in centimeters (value bounds must be between 30 and 250)

PatientGender

Patient’s gender (0 = male, 1 = female)

PeakPressureAlarmThreshold

Threshold for peak pressure alarm in mmH2O (value bounds must be between 50 and 700)

EolConfirm

Confirm end-of-line test step (value bounds must be between 0 and 0)

Implementations

impl ControlSetting

pub fn default(&self) -> usize

Default settings

pub fn bounds(&self) -> RangeInclusive<usize>

Allowed value bounds per setting