Function makair_telemetry::gather_telemetry_from_bytes

pub fn gather_telemetry_from_bytes(
    telemetry_bytes_rx: Receiver<Vec<u8>>, 
    telemetry_tx: Sender<TelemetryChannelType>, 
    control_rx: Option<Receiver<ControlMessage>>, 
    control_bytes_tx: Option<Sender<Vec<u8>>>, 
    sleep_duration: Option<Duration>
) -> !

Open a byte channel, consume it endlessly and send parsed telemetry messages through another channel

telemetry_bytes_rx - Receiver of a channel used to transport telemetry bytes (input).

telemetry_tx - Sender of a channel used to transport structured telemetry messages (output).

control_rx - Optional receiver of a channel used to transport structured control messages (input).

control_bytes_tx - Optional sender of a channel used to transport control bytes (output).

sleep_duration - Optional duration to wait when there are no more bytes to parse; if None then no sleep.

This is meant to be run in a dedicated thread.