Function makair_telemetry::gather_telemetry_from_file

pub fn gather_telemetry_from_file(
    file: File, 
    tx: Sender<TelemetryChannelType>, 
    enable_time_simulation: bool
)

Open a file containing serialized telemetry data, read it and send back parsed telemetry messages through a channel

file - Handle to a file that contains telemetry data.

tx - Sender of a channel.

enable_time_simulation - If true, telemetry messages will be sent in a realistic timing; if false, they will be read as fast as possible.

This is meant to be run in a dedicated thread.